Restructure tests layout to match project modules
