+++ setup.cfg
@@ -11,7 +11,7 @@ testdocker       = pytest --addopts "-m 'dockertest'"
 
 # Component level filtering of tests
 testbuilder      = pytest --addopts "-k builder"
-testdataprovider = pytest --addopts "-k provider"
+testdataprovider = pytest --addopts "-k data_provider"
 testdataset      = pytest --addopts "-k dataset"
 testmodel        = pytest --addopts "-k model"
 testserializer   = pytest --addopts "-k serializer"
@@ -19,6 +19,8 @@ testserver       = pytest --addopts "-k server"
 testwatchman     = pytest --addopts "-k watchman"
 testcli          = pytest --addopts "-k cli"
 testclient       = pytest --addopts "-k client"
+testutil         = pytest --addopts "-k util"
+testexamples     = pytest --addopts "-k examples"
 
 # Black formatting
 testformatting   = pytest --addopts "-k formatting"
