+++ tests/test_data_provider/test_azure_utils.py
