+++ tests/test_model/test_model.py
@@ -5,6 +5,7 @@ import tempfile
 import logging
 import pydoc
 
+import pytest
 import numpy as np
 
 from sklearn.exceptions import NotFittedError
