+++ tests/test_server/test_gordo_server.py
