+++ tests/test_util/test_disk_registry.py
